carts, orders, payments 모듈 로그 추가

+++ src/orders/carts.service.ts
@@ -8,6 +8,9 @@ import { Users } from '../user/entities/user.entitiy';
 import { Point } from '../point/entities/point.entity';
 import { Payments } from '../payments/entities/payments.entity';
 import { Orders } from './entities/orders.entity';
+import _ from 'lodash';
+import logger from 'src/common/log/logger';
+import { validation } from 'src/common/pipe/validationPipe';
 
 @Injectable()
 export class CartService {
@@ -28,6 +31,15 @@ export class CartService {
     goodsId: number,
     createCartDto: CreateCartDto,
   ) {
+
+    await validation(CreateCartDto, createCartDto)
+
+    if (_.isNil(userId) || userId == 0 || _.isNil(goodsId) || goodsId == 0) {
+      const error = new BadRequestException('잘못된 요청입니다!') 
+      logger.errorLogger(error, `userId = ${userId}, goodsId = ${goodsId}, createCartDto = ${JSON.stringify(createCartDto)}`)
+      throw error
+    }
+
     const { ctCount } = createCartDto;
 
     const goods = await this.goodsRepository.findOne({
@@ -37,13 +49,17 @@ export class CartService {
       },
     });
     if (!goods) {
-      throw new BadRequestException('존재하지 않는 상품입니다.');
+      const error = new BadRequestException('존재하지 않는 상품입니다.'); 
+      logger.errorLogger(error, `userId = ${userId}, goodsId = ${goodsId}, createCartDto = ${JSON.stringify(createCartDto)}, goods = ${goods}`)
+      throw error
     }
 
     const count = goods.stock.count - ctCount;
 
     if (count < 0) {
-      throw new BadRequestException('재고가 없습니다.');
+      const error = new BadRequestException('재고가 없습니다.'); 
+      logger.errorLogger(error, `userId = ${userId}, goodsId = ${goodsId}, createCartDto = ${JSON.stringify(createCartDto)}`)
+      throw error
     }
 
 
@@ -53,7 +69,6 @@ export class CartService {
         goods_id: goodsId
       },
     });
-    console.log('국밥!!!!!!!!', existingCartItem, userId, goodsId)
 
     if (existingCartItem) {
       // 이미 장바구니에 있는 상품인 경우 수량만 업데이트
@@ -67,32 +82,30 @@ export class CartService {
       ct_count: ctCount,
       ct_price: goods.g_price,
     });
-    return this.cartsRepository.save(newCartItem);
-
+    return await this.cartsRepository.save(newCartItem);
 
   }
 
 
   // 장바구니 특정 상품 삭제(cartId를 goodsId로 바꾸는게 나을 거 같은데<-완료)
   async removeFromCart(userId: number, goodsId: number) {
-    if (!goodsId || goodsId == 0) {
-      throw new BadRequestException('잘못된 요청입니다!')
+    if (_.isNil(userId) || userId == 0 || _.isNil(goodsId) || goodsId == 0) {
+      const error = new BadRequestException('잘못된 요청입니다!') 
+      logger.errorLogger(error, `userId = ${userId}, goodsId = ${goodsId}`)
+      throw error
     }
-
-    // if(!goodsId || goodsId == 0) {
-    //   throw new BadRequestException('잘못된 요청입니다!')
-    // }
 
     const cartInfo = await this.cartsRepository.findOne({
       where: {
         goods_id: goodsId,
         id: userId
-        // goods_id: goodsId
       }
     })
 
     if (!cartInfo) {
-      throw new NotFoundException('장바구니에 해당 상품이 존재하지 않습니다.')
+      const  error = new NotFoundException('장바구니에 해당 상품이 존재하지 않습니다.');
+      logger.errorLogger(error, `userId = ${userId}, goodsId = ${goodsId}, cartInfo = ${JSON.stringify(cartInfo)}`)
+      throw error
     }
 
     await this.cartsRepository.delete(
@@ -109,12 +122,27 @@ export class CartService {
 
   // 장바구니 특정 상품 수량 변경
   async updateQuantity(userId: number, cartId: number, ctCounts: number): Promise<Carts> {
+
+    if (_.isNil(userId) || userId == 0 || _.isNil(cartId) || cartId == 0 || _.isNil(ctCounts) || ctCounts == 0) {
+      const error = new BadRequestException('잘못된 요청입니다!') 
+      logger.errorLogger(error, `userId = ${userId}, cartId = ${cartId}`)
+      throw error
+    }
+
+    if (_.isNil(ctCounts) || ctCounts == 0) {
+      const error = new BadRequestException('변경할 수량을 입력해 주세요.') 
+      logger.errorLogger(error, `userId = ${userId}, cartId = ${cartId}, ctCounts = ${ctCounts}`)
+      throw error
+    }
+
     const cartItem = await this.cartsRepository.findOne({
       where: { id: cartId },
     });
 
     if (!cartItem) {
-      throw new Error('장바구니에 상품을 찾을 수 없습니다.');
+      const  error = new NotFoundException('장바구니에 상품을 찾을 수 없습니다.');
+      logger.errorLogger(error, `userId = ${userId}, cartId = ${cartId}, cartItem = ${JSON.stringify(cartItem)}`)
+      throw error
     }
 
     cartItem.ct_count = ctCounts;
@@ -122,7 +150,9 @@ export class CartService {
       where: { id: cartItem.goods_id },
     });
     if (!goods) {
-      throw new Error('상품을 찾을 수 없습니다.');
+      const  error = new NotFoundException('해당 상품을 찾을 수 없습니다.');
+      logger.errorLogger(error, `userId = ${userId}, cartId = ${cartId}, cartItem = ${JSON.stringify(cartItem)}, goods = ${goods}`)
+      throw error
     }
     cartItem.ct_price = ctCounts * goods.g_price; // 상품 가격과 수량을 곱하여 총 가격 업데이트
 
@@ -131,14 +161,28 @@ export class CartService {
 
   //소비자 장바구니의 모든 상품 조회
   async getCartItems(userId: number): Promise<Carts[]> {
-    return this.cartsRepository.find({
+
+    if (_.isNil(userId) || userId == 0) {
+      const error = new BadRequestException('잘못된 요청입니다!') 
+      logger.errorLogger(error, `userId = ${userId}`)
+      throw error
+    }
+
+    const carts = await this.cartsRepository.find({
       where: { id: userId, },
     });
+
+    return carts
   }
 
 
   //장바구니 비우기
   async clearCart(userId: number): Promise<void> {
+    if (_.isNil(userId) || userId == 0) {
+      const error = new BadRequestException('잘못된 요청입니다!') 
+      logger.errorLogger(error, `userId = ${userId}`)
+      throw error
+    }
     await this.cartsRepository.delete({ id: userId });
   }
 }
